Update Flight POST and DELETE endpoints

+++ src/main/java/com/alikemal/flightbooking/flight/FlightController.java
@@ -1,5 +1,10 @@
 package com.alikemal.flightbooking.flight;
 
+import com.alikemal.flightbooking.aircraft.Aircraft;
+import com.alikemal.flightbooking.aircraft.AircraftService;
+import com.alikemal.flightbooking.airport.Airport;
+import com.alikemal.flightbooking.airport.AirportService;
+import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -7,12 +12,11 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("api/v1/flights")
+@AllArgsConstructor
 public class FlightController {
     private final FlightService flightService;
-
-    public FlightController(FlightService flightService) {
-        this.flightService = flightService;
-    }
+    private final AirportService airportService;
+    private final AircraftService aircraftService;
 
     @GetMapping("/{flightNumber}")
     public Flight getFlightByFlightNumber(@PathVariable Long flightNumber) {
@@ -26,11 +30,41 @@ public class FlightController {
 
     @DeleteMapping("/{flightNumber}")
     public void deleteFlightByFlightNumber(@PathVariable Long flightNumber) {
+        // Check if flight exists
+        Flight flight = flightService.getFlightByFlightNumber(flightNumber);
+
+        if (flight == null) {
+            throw new IllegalArgumentException("Flight not found with flight number: " + flightNumber);
+        }
+
         flightService.deleteFlightByFlightNumber(flightNumber);
     }
 
     @PostMapping
-    public void addFlight(@RequestBody Flight flight) {
+    public void addFlight(@RequestBody FlightPostRequest flightPost) {
+        // Get departure and destination airports
+        Airport departure = airportService.getAirportByCode(flightPost.getDeparture());
+        Airport destination = airportService.getAirportByCode(flightPost.getDestination());
+
+        // Get Aircraft
+        Aircraft aircraft = aircraftService.getAircraftByCode(flightPost.getAircraft());
+
+        // Create a new flight
+        Flight flight = new Flight();
+        flight.setDeparture(departure);
+        flight.setDestination(destination);
+        flight.setDepartureTime(flightPost.getDepartureTime());
+        flight.setArrivalTime(flightPost.getArrivalTime());
+        flight.setPrice(flightPost.getPrice());
+        flight.setAvailableTickets(flightPost.getAvailableTickets());
+        flight.setAircraft(aircraft);
+        flight.setStatus(flightPost.getStatus());
+
+        // Check if flight contains null attribute
+        if (flight.getDepartureTime() == null || flight.getArrivalTime() == null || flight.getPrice() == null || flight.getAvailableTickets() == null) {
+            throw new IllegalArgumentException("Flight cannot contain null attributes");
+        }
+
         flightService.addFlight(flight);
     }
